requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
blas=1.0=mkl
boost=1.74.0=py38h2b96118_5
boost-cpp=1.74.0=h359cf19_6
bottleneck=1.3.7=py38ha9d4c09_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
cairo=1.16.0=h19f5f5c_2
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.5=py38hdb19cb5_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
expat=2.6.3=h6a678d5_0
fcd_torch=1.0.7=pyh9f0ad1d_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fontconfig=2.14.0=h8e229c2_0
fonttools=4.51.0=py38h5eee18b_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.5.0=pypi_0
fsspec=2024.10.0=pypi_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
greenlet=3.0.1=py38h6a678d5_0
icu=69.1=h9c3ff4c_0
idna=3.7=py38h06a4308_0
importlib_resources=6.4.0=py38h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py38h06a4308_0
joblib=1.4.2=py38h06a4308_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.4=py38h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.11.1.6=0
libcurand=10.3.7.77=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py38h5eee18b_0
matplotlib-base=3.7.2=py38h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
ml-collections=0.1.1=pypi_0
molsets=0.3.1=py_0
moses=0.10.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=py38h06a4308_0
pandas=1.4.4=pypi_0
pcre2=10.42=hebb0a14_1
pillow=10.4.0=py38h5eee18b_0
pip=24.2=py38h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=3.10.0=py38h06a4308_0
pomegranate=0.12.0=py38hc9558a2_0
pooch=1.7.0=py38h06a4308_0
propcache=0.2.0=pypi_0
psutil=6.1.0=pypi_0
pycairo=1.23.0=py38hd1222b9_0
pyparsing=3.1.4=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.20=he870216_0
python-dateutil=2.9.0post0=py38h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py38h06a4308_0
pyyaml=6.0.2=py38h5eee18b_0
rdkit=2022.3.5=pypi_0
rdkit-pypi=2022.9.5=pypi_0
readline=8.2=h5eee18b_0
reportlab=3.5.67=py38hfdd840d_1
requests=2.32.3=py38h06a4308_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=py38h06a4308_0
setuptools=75.1.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlalchemy=2.0.34=py38h00e1ef3_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.2=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
torch-cluster=1.6.3+pt20cu117=pypi_0
torch-geometric=2.1.0.post1=pypi_0
torch-scatter=2.1.2+pt20cu117=pypi_0
torch-sparse=0.6.18+pt20cu117=pypi_0
torch-spline-conv=1.2.2+pt20cu117=pypi_0
torchaudio=2.0.2=py38_cu117
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu117
tqdm=4.67.0=pypi_0
typing-extensions=4.11.0=py38h06a4308_0
typing_extensions=4.11.0=py38h06a4308_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.2.3=py38h06a4308_0
wheel=0.44.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.15.2=pypi_0
zipp=3.20.2=py38h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
